+++ packages/stack-shared/src/interface/crud/projects.ts
@@ -8,21 +8,14 @@ const teamPermissionSchema = yupObject({
 
 const oauthProviderSchema = yupObject({
   id: schemaFields.oauthIdSchema.defined(),
-  enabled: schemaFields.oauthEnabledSchema.defined(),
   type: schemaFields.oauthTypeSchema.defined(),
   client_id: schemaFields.yupDefinedAndNonEmptyWhen(
     schemaFields.oauthClientIdSchema,
-    {
-      type: 'standard',
-      enabled: true,
-    },
+    { type: 'standard' },
   ),
   client_secret: schemaFields.yupDefinedAndNonEmptyWhen(
     schemaFields.oauthClientSecretSchema,
-    {
-      type: 'standard',
-      enabled: true,
-    },
+    { type: 'standard' },
   ),
 
   // extra params
@@ -72,8 +65,8 @@ export const projectsCrudAdminReadSchema = yupObject({
   created_at_millis: schemaFields.projectCreatedAtMillisSchema.defined(),
   user_count: schemaFields.projectUserCountSchema.defined(),
   is_production_mode: schemaFields.projectIsProductionModeSchema.defined(),
+  /** @deprecated */
   config: yupObject({
-    id: schemaFields.projectConfigIdSchema.defined(),
     allow_localhost: schemaFields.projectAllowLocalhostSchema.defined(),
     sign_up_enabled: schemaFields.projectSignUpEnabledSchema.defined(),
     credential_enabled: schemaFields.projectCredentialEnabledSchema.defined(),
@@ -93,7 +86,7 @@ export const projectsCrudAdminReadSchema = yupObject({
     team_member_default_permissions: yupArray(teamPermissionSchema.defined()).defined(),
     user_default_permissions: yupArray(teamPermissionSchema.defined()).defined(),
     oauth_account_merge_strategy: schemaFields.oauthAccountMergeStrategySchema.defined(),
-  }).defined(),
+  }).defined().meta({ openapiField: { hidden: true } }),
 }).defined();
 
 export const projectsCrudClientReadSchema = yupObject({
@@ -109,7 +102,7 @@ export const projectsCrudClientReadSchema = yupObject({
     allow_user_api_keys: schemaFields.yupBoolean().defined(),
     allow_team_api_keys: schemaFields.yupBoolean().defined(),
     enabled_oauth_providers: yupArray(enabledOAuthProviderSchema.defined()).defined().meta({ openapiField: { hidden: true } }),
-  }).defined(),
+  }).defined().meta({ openapiField: { hidden: true } }),
 }).defined();