+++ packages/stack-shared/src/known-errors.tsx
@@ -635,6 +635,20 @@ const ProjectNotFound = createKnownErrorConstructor(
   (json: any) => [json.project_id] as const,
 );
 
+const BranchDoesNotExist = createKnownErrorConstructor(
+  KnownError,
+  "BRANCH_DOES_NOT_EXIST",
+  (branchId: string) => [
+    400,
+    `The branch with ID ${branchId} does not exist.`,
+    {
+      branch_id: branchId,
+    },
+  ] as const,
+  (json: any) => [json.branch_id] as const,
+);
+
+
 const SignUpNotEnabled = createKnownErrorConstructor(
   KnownError,
   "SIGN_UP_NOT_ENABLED",
@@ -885,6 +899,21 @@ const PermissionNotFound = createKnownErrorConstructor(
   (json: any) => [json.permission_id] as const,
 );
 
+const PermissionScopeMismatch = createKnownErrorConstructor(
+  KnownError,
+  "WRONG_PERMISSION_SCOPE",
+  (permissionId: string, expectedScope: "team" | "project", actualScope: "team" | "project" | null) => [
+    404,
+    `Permission ${JSON.stringify(permissionId)} not found. (It was found for a different scope ${JSON.stringify(actualScope)}, but scope ${JSON.stringify(expectedScope)} was expected.)`,
+    {
+      permission_id: permissionId,
+      expected_scope: expectedScope,
+      actual_scope: actualScope,
+    },
+  ] as const,
+  (json: any) => [json.permission_id, json.expected_scope, json.actual_scope] as const,
+);
+
 const ContainedPermissionNotFound = createKnownErrorConstructor(
   KnownError,
   "CONTAINED_PERMISSION_NOT_FOUND",
@@ -1004,7 +1033,7 @@ const InvalidOAuthClientIdOrSecret = createKnownErrorConstructor(
   "INVALID_OAUTH_CLIENT_ID_OR_SECRET",
   (clientId?: string) => [
     400,
-    "The OAuth client ID or secret is invalid. The client ID must be equal to the project ID, and the client secret must be a publishable client key.",
+    "The OAuth client ID or secret is invalid. The client ID must be equal to the project ID (potentially with a hash and a branch ID), and the client secret must be a publishable client key.",
     {
       client_id: clientId ?? null,
     },
@@ -1365,6 +1394,7 @@ export const KnownErrors = {
   ApiKeyNotFound,
   PublicApiKeyCannotBeRevoked,
   ProjectNotFound,
+  BranchDoesNotExist,
   SignUpNotEnabled,
   PasswordAuthenticationNotEnabled,
   PasskeyAuthenticationNotEnabled,
@@ -1388,6 +1418,7 @@ export const KnownErrors = {
   PasskeyWebAuthnError,
   PasskeyAuthenticationFailed,
   PermissionNotFound,
+  PermissionScopeMismatch,
   ContainedPermissionNotFound,
   TeamNotFound,
   TeamMembershipNotFound,