+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,37 @@
 # react-example
 
+## 2.8.11
+
+### Patch Changes
+
+- @stackframe/react@2.8.11
+
+## 2.8.10
+
+### Patch Changes
+
+- @stackframe/react@2.8.10
+
+## 2.8.9
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/react@2.8.9
+
+## 2.8.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/react@2.8.8
+
+## 2.8.7
+
+### Patch Changes
+
+- @stackframe/react@2.8.7
+
 ## 2.8.6
 
 ### Patch Changes